+++ compiler-core/src/ast/visit.rs
@@ -43,7 +43,8 @@ use crate::{
     ast::{
         BitArraySize, RecordBeingUpdated, TypeAstConstructorName, TypedBitArraySize,
         TypedConstantBitArraySegment, TypedDefinitions, TypedImport, TypedTailPattern,
-        TypedTypeAlias, typed::InvalidExpression,
+        TypedTypeAlias,
+        typed::{InvalidExpression, RecordUpdateAssignment},
     },
     exhaustiveness::CompiledCase,
     parse::LiteralFloatValue,
@@ -330,7 +331,7 @@ pub trait Visit<'ast> {
         &mut self,
         location: &'ast SrcSpan,
         type_: &'ast Arc<Type>,
-        record: &'ast Option<Box<TypedAssignment>>,
+        record: &'ast Option<Box<RecordUpdateAssignment>>,
         constructor: &'ast TypedExpr,
         arguments: &'ast [TypedCallArg],
     ) {
@@ -1673,7 +1674,7 @@ pub fn visit_typed_expr_record_update<'a, V>(
     v: &mut V,
     _location: &'a SrcSpan,
     _type_: &'a Arc<Type>,
-    record: &'a Option<Box<TypedAssignment>>,
+    record: &'a Option<Box<RecordUpdateAssignment>>,
     constructor: &'a TypedExpr,
     arguments: &'a [TypedCallArg],
 ) where
@@ -1681,7 +1682,7 @@ pub fn visit_typed_expr_record_update<'a, V>(
 {
     v.visit_typed_expr(constructor);
     if let Some(record) = record {
-        v.visit_typed_assignment(record);
+        v.visit_typed_expr(&record.value);
     }
     for argument in arguments {
         v.visit_typed_call_arg(argument);